XIVY-4974 refactor: split pilot tests and re-use db fixture provisioning
