+++ database/db-demos-test/src_test/db/demo/persistence/test/PersonRepoIvyTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
 
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import ch.ivyteam.ivy.environment.IvyTest;
@@ -17,102 +19,58 @@ import jakarta.data.page.PageRequest;
 class PersonRepoIvyTest {
 
   PersonRepo repository = Person.repository();
-
-  @Test
-  void useRepo() {
-    var all = repository.findAll().toList();
-    assertThat(all).isEmpty();
-
-    var james = new Person();
-    james.setFirstName("James");
-    james.setLastName("Bond");
-    james.setBirthDate(new Date(1968, 4, 13));
-
-    repository.insert(james);
-    assertThat(repository.findAll().toList())
-        .hasSize(1);
-    var persisted = repository.findAll()
-        .filter(p -> "James".equals(p.getFirstName()))
-        .findFirst();
-    assertThat(persisted).isPresent();
-    assertThat(persisted.get().getId())
-        .isNotNull();
-
-    repository.delete(persisted.get());
-    assertThat(repository.findAll().toList())
-        .isEmpty();
-  }
+  List<Person> pilots;
 
   @Test
   @SuppressWarnings("unchecked")
   void query_findByLastName() {
-    List<Person> pilots = AviationFactory.createPilots();
-    repository.insertAll(pilots);
+    List<Person> wrightBrothers = repository.findByLastName("Wright",
+        _Person.birthDate.asc(),
+        _Person.firstName.descIgnoreCase());
+    assertThat(wrightBrothers)
+        .as("Optional sorting parameters create ordered results")
+        .extracting(Person::getFirstName)
+        .containsExactly("Wilbur", "Orville");
 
-    try {
-      List<Person> wrightBrothers = repository.findByLastName("Wright",
-          _Person.birthDate.asc(),
-          _Person.firstName.descIgnoreCase());
-      assertThat(wrightBrothers)
-          .as("Optional sorting parameters create ordered results")
-          .extracting(Person::getFirstName)
-          .containsExactly("Wilbur", "Orville");
-
-      assertThat(repository.findByLastName("Blériot"))
-          .extracting(Person::getFirstName)
-          .containsOnly("Louis Charles Joseph");
-    } finally {
-      repository.deleteAll(pilots);
-    }
+    assertThat(repository.findByLastName("Blériot"))
+        .extracting(Person::getFirstName)
+        .containsOnly("Louis Charles Joseph");
   }
 
   @Test
   void query_findByLikePattern() {
-    List<Person> pilots = AviationFactory.createPilots();
-    repository.insertAll(pilots);
-
-    try {
-      assertThat(repository.findByNamePart("Wri%"))
-          .extracting(Person::getLastName)
-          .containsOnly("Wright");
-
-    } finally {
-      repository.deleteAll(pilots);
-    }
+    assertThat(repository.findByNamePart("Wri%"))
+        .extracting(Person::getLastName)
+        .containsOnly("Wright");
   }
 
   @Test
   void query_byJDQL() {
-    List<Person> pilots = AviationFactory.createPilots();
-    repository.insertAll(pilots);
-
-    try {
-      var the1867s = repository.bornIn(1867);
-      assertThat(the1867s)
-          .as("JDQL query statements empower repositories to do advanced search queries")
-          .extracting(Person::getFirstName)
-          .containsOnly("Wilbur");
-
-    } finally {
-      repository.deleteAll(pilots);
-    }
+    var the1867s = repository.bornIn(1867);
+    assertThat(the1867s)
+        .as("JDQL query statements empower repositories to do advanced search queries")
+        .extracting(Person::getFirstName)
+        .containsOnly("Wilbur");
   }
 
   @Test
   void query_distinctNamesWithFilters() {
-    List<Person> pilots = AviationFactory.createPilots();
+    var names = repository.namesOfLength(3, 10, PageRequest.ofSize(10));
+    assertThat(names)
+        .as("omit names longer than 10 characters")
+        .containsOnly("Wilbur", "Orville")
+        .doesNotContain("Louis Charles Joseph");
+  }
+
+  @BeforeEach
+  void setUp() {
+    pilots = AviationFactory.createPilots();
     repository.insertAll(pilots);
+  }
 
-    try {
-      var names = repository.namesOfLength(3, 10, PageRequest.ofSize(10));
-      assertThat(names)
-          .as("omit names longer than 10 characters")
-          .containsOnly("Wilbur", "Orville")
-          .doesNotContain("Louis Charles Joseph");
-
-    } finally {
-      repository.deleteAll(pilots);
-    }
+  @AfterEach
+  void tearDown() {
+    repository.deleteAll(pilots);
   }
 
   static class AviationFactory {
